producer/requirements.txt
websocket-client
confluent_kafka
